Add Docker support for the ecu_dash project; Configure static files and Nginx; Refactor course listing and faculty filtering functionality
This commit introduces Docker configurations to the project, optimizing the development and deployment process. Files like '.dockerignore', 'compose.yaml', and 'Dockerfile' have been added to define the Docker build context, set up multi-container applications, and describe the build instructions for the Docker image, respectively. It also includes a 'README.Docker.md' guide for building and deploying the application using Docker commands.

Revised the course listing and faculty filtering code in the views and template files. This included improvements such as adding sort functionality to the faculty list and updating the course filtering logic to allow for more flexible search. This refactoring enhances the user experience and simplifies future code modifications.

In the settings.py file, the STATIC_ROOT setting was added to point to a new directory for static files. A new Nginx configuration file was also created to serve static files and set the proxy pass to the web server.